+++ src/platforms/android/androidiaphandler.h
@@ -5,7 +5,7 @@
 #ifndef ANDROIDIAPHANDLER_H
 #define ANDROIDIAPHANDLER_H
 
-#include "iaphandler.h"
+#include "purchasehandler.h"
 #include "jni.h"
 
 class AndroidIAPHandler final : public IAPHandler {
