+++ src/purchasehandler.cpp
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "iaphandler.h"
+#include "purchasehandler.h"
 #include "constants.h"
 #include "inspector/inspectorhandler.h"
 #include "leakdetector.h"
